+++ UnityProject/Assets/OpenMined.Tests/Editor/FloatTensor/FloatTensorTest.cs
@@ -1,7 +1,6 @@
 using System;
 using NUnit.Framework;
 using OpenMined.Network.Controllers;
-using OpenMined.Syft.NN;
 
 namespace OpenMined.Tests.Editor.FloatTensor
 {
@@ -1860,7 +1859,7 @@ namespace OpenMined.Tests.Editor.FloatTensor
 
             var tensor = ctrl.floatTensorFactory.Create(_data: data, _shape: shape);
 
-            var actualTensor = Functional.Softmax(tensor);
+            var actualTensor = tensor.Softmax();
 
             var expectedTensor = new float[]
                 {(float) 0.0320586, (float) 0.08714432, (float) 0.23688282, (float) 0.64391426};
@@ -1880,16 +1879,10 @@ namespace OpenMined.Tests.Editor.FloatTensor
             var tensor = ctrl.floatTensorFactory.Create(_data: data, _shape: shape, _autograd:true);
             var gradTensor = ctrl.floatTensorFactory.Create(_data: gradData, _shape: shape);
 
-            var outputTensor = Functional.Softmax(tensor);
-            
-            var gradInput = Functional.SoftmaxGradient(outputTensor, gradTensor, 0);
+            var outputTensor = tensor.Softmax();
 
             var expectedTensor = new float[]
                 {(float) 0.2280, (float) -0.0916, (float) -0.0750, (float) -0.0614};
-            for (var i = 0; i < expectedTensor.Length; i++)
-            {
-                Assert.AreEqual(expectedTensor[i], gradInput[i], 1e-3);
-            }
             
             outputTensor.Backward(gradTensor, null);
             for (var i = 0; i < expectedTensor.Length; i++)
@@ -1906,7 +1899,7 @@ namespace OpenMined.Tests.Editor.FloatTensor
 
             var tensor = ctrl.floatTensorFactory.Create(_data: data, _shape: shape);
 
-            var actualTensor = Functional.Softmax(tensor);
+            var actualTensor = tensor.Softmax();
 
             var expectedData = new float[] {(float) 0.2689, (float) 0.7311, (float) 0.2689, (float) 0.7311};
             var expectedTensor = ctrl.floatTensorFactory.Create(_data: expectedData, _shape: shape);
@@ -1915,7 +1908,7 @@ namespace OpenMined.Tests.Editor.FloatTensor
                 Assert.AreEqual(expectedTensor[i], actualTensor[i], 1e-3);
             }
 
-            actualTensor = Functional.Softmax(tensor, 0);
+            actualTensor = tensor.Softmax(0);
 
             expectedData = new float[] {(float) 0.1192, (float) 0.1192, (float) 0.8808, (float) 0.8808};
             expectedTensor = ctrl.floatTensorFactory.Create(_data: expectedData, _shape: shape);
@@ -1938,7 +1931,7 @@ namespace OpenMined.Tests.Editor.FloatTensor
             };
 
             var tensor = ctrl.floatTensorFactory.Create(_data: data, _shape: shape);
-            var actualTensor = Functional.Softmax(tensor);
+            var actualTensor = tensor.Softmax();
 
             var expectedTensor = ctrl.floatTensorFactory.Create(_data: expectedData, _shape: shape);
             for (var i = 0; i < expectedTensor.Size; i++)
@@ -1946,7 +1939,7 @@ namespace OpenMined.Tests.Editor.FloatTensor
                 Assert.AreEqual(expectedTensor[i], actualTensor[i], 1e-3);
             }
 
-            actualTensor = Functional.Softmax(tensor, 1);
+            actualTensor = tensor.Softmax(1);
             expectedData = new float[]
             {
                 (float) 0.1192, (float) 0.1192, (float) 0.8808, (float) 0.8808, (float) 0.1192, (float) 0.1192,
@@ -1958,7 +1951,7 @@ namespace OpenMined.Tests.Editor.FloatTensor
                 Assert.AreEqual(expectedTensor[i], actualTensor[i], 1e-3);
             }
             
-            actualTensor = Functional.Softmax(tensor, 0);
+            actualTensor = tensor.Softmax(0);
             expectedData = new float[]
             {
                 (float) 0.0180, (float) 0.0180, (float) 0.0180, (float) 0.0180, (float) 0.9820, (float) 0.9820,
