--- UnityProject/Assets/OpenMined/Syft/Model.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 327d6b2bd20e34e73b955dccd260df7a
-folderAsset: yes
-timeCreated: 1512723902
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
