+++ UnityProject/Assets/OpenMined/Syft/NN.meta
@@ -1,3 +1,10 @@
-﻿fileFormatVersion: 2
-guid: 4b5baf45fd014a699f229f43d95725a8
-timeCreated: 1513099648
+fileFormatVersion: 2
+guid: 327d6b2bd20e34e73b955dccd260df7a
+folderAsset: yes
+timeCreated: 1512723902
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
