--- UnityProject/Assets/OpenMined/Syft/NN/Functional.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 4b746744390f4cbe89f61546b9def3a5
-timeCreated: 1513099741
