+++ UnityProject/Assets/OpenMined/Syft/NN/Sigmoid.cs
