+++ UnityProject/Assets/OpenMined/Syft/NN/Tanh.cs
